Revert "Feat/create user (#5)" (#6)
This reverts commit 952d2894acbe12a0e97b291223b34adebfd49d50.

+++ models/transfer.go
@@ -3,7 +3,7 @@ package models
 // Transfers entity struct.
 type Transfer struct {
 	Model
-	FromAccountID uint  `gorm:"not null;index;index:from_to_accounts_idx" json:"from_account_id"`
-	ToAccountID   uint  `gorm:"not null;index;index:from_to_accounts_idx" json:"to_account_id"`
+	FromAccountID int64 `gorm:"not null;index;index:from_to_accounts_idx" json:"from_account_id"`
+	ToAccountID   int64 `gorm:"not null;index;index:from_to_accounts_idx" json:"to_account_id"`
 	Amount        int64 `gorm:"not null;comment:must be positive" json:"amount"`
 }
